refactor: 패키지 구조 및 검증로직 추가
