+++ src/main/java/domain/LottoManager.java
@@ -1,13 +1,22 @@
-package lotto;
+package domain;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
+
+import domain.amount.LottoMoney;
+import domain.lotto.Lotto;
+import domain.lotto.WinningLotto;
+import domain.statistics.Rank;
 
 public class LottoManager {
 	private final List<Lotto> lottos;
 	private final WinningLotto winningLotto;
 
 	public LottoManager(List<Lotto> lottos, WinningLotto winningLotto) {
+		Objects.requireNonNull(lottos);
+		Objects.requireNonNull(winningLotto);
+
 		this.lottos = lottos;
 		this.winningLotto = winningLotto;
 	}
@@ -17,6 +26,9 @@ public class LottoManager {
 	}
 
 	public double calculateProfit(Map<Rank, Integer> countByRank, LottoMoney lottoMoney) {
+		Objects.requireNonNull(countByRank);
+		Objects.requireNonNull(lottoMoney);
+
 		long winningMoney = 0L;
 
 		for (Rank rank : Rank.values()) {
