+++ src/main/java/domain/lotto/WinningLotto.java
@@ -1,18 +1,27 @@
-package lotto;
+package domain.lotto;
 
-import lotto.number.LottoNumber;
+import java.util.Objects;
+
+import domain.lotto.number.LottoNumber;
 
 public class WinningLotto {
 	private final Lotto lotto;
 	private final LottoNumber bonusNumber;
 
 	public WinningLotto(Lotto lotto, LottoNumber lottoNumber) {
-		if (lotto.isContains(lottoNumber)) {
-			throw new IllegalArgumentException(
-				"당첨번호와 보너스번호는 중복될 수 없습니다. lotto=" + lotto + "lottoNumber" + lottoNumber);
-		}
+		validate(lotto, lottoNumber);
 		this.lotto = lotto;
 		this.bonusNumber = lottoNumber;
+	}
+
+	private void validate(Lotto lotto, LottoNumber lottoNumber) {
+		Objects.requireNonNull(lotto);
+		Objects.requireNonNull(lottoNumber);
+
+		if (lotto.isContains(lottoNumber)) {
+			throw new IllegalArgumentException(
+				"당첨번호와 보너스번호는 중복될 수 없습니다. domain.lotto=" + lotto + "lottoNumber" + lottoNumber);
+		}
 	}
 
 	public int countMatch(Lotto lotto) {
@@ -26,7 +35,7 @@ public class WinningLotto {
 	@Override
 	public String toString() {
 		return "WinningLotto{" +
-			"lotto=" + lotto +
+			"domain.lotto=" + lotto +
 			", bonusNumber=" + bonusNumber +
 			'}';
 	}
