+++ src/test/java/domain/amount/element/ManualLottoAmountTest.java
@@ -1,11 +1,11 @@
-package lotto.amount.element;
+package domain.amount.element;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import lotto.LottoMoney;
+import domain.amount.LottoMoney;
 
 class ManualLottoAmountTest {
 	@DisplayName("수동 로또 생성시, 수량이 금액보다 많으면 에러가 발생한다.")
