+++ src/test/java/domain/lotto/number/LottoNumberTest.java
@@ -1,4 +1,4 @@
-package lotto.number;
+package domain.lotto.number;
 
 import static org.assertj.core.api.Assertions.*;
